renamed solver and separated it into more specialized classes
